# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0"
  title: KIWAME API REST
  description: API for the management of KIWAME 
...
securityDefinitions:
  ApiKey:
    type: "apiKey"
    name: "x-api-key"
    in: "header"
    description: "Security sheme migrated from security policy"
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
     For accessing the API a valid JWT token must be passed in all the queries in
     the 'Authorization' header.


     A valid JWT token is generated by the API and retourned as answer of a call
     to the route /login giving a valid user & password.


     The following syntax must be used in the 'Authorization' header :

       Bearer xxxxxx.yyyyyyy.zzzzzz


# during dev, should point to your local machine
#host: 80.58.61.69  #PRODUCCIÓN
#host: 138.4.47.33:8163  #UPM DEV
#host: localhost:8090  #LOCAL DEV
# basePath prefixes all resource paths
basePath: / #dds
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /rule/{id}:
    x-swagger-router-controller: rule
    get:
      tags:
        - rules
      summary: Toogle enabled 
      description: toogle enable/disabled rule
      operationId: findById
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - admin
      consumes:
        - application/json
        - application/octet-stream
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the rule to toogle
          type: string
          required: true
      responses:
        "200":
          description: rule retrieves
          schema:
            $ref: "#/definitions/RuleResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Rule Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
    put:
      tags:
        - rules
      summary: Toogle enabled 
      description: toogle enable/disabled rule
      operationId: toogleEnabled
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - admin
      consumes:
        - application/json
        - application/octet-stream
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the rule to toogle
          type: string
          required: true
      responses:
        "200":
          description: rule Enabled/Disabled
          schema:
            $ref: "#/definitions/ServerResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Rule Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
    delete:
      tags:
        - rules
      summary: Delete rule
      description: Delete rule
      operationId: deleteRule
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - admin
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the rule to delete
          type: string
          required: true
      responses:
        "200":
          description: Rule removed
          schema:
            $ref: "#/definitions/ServerResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Profile Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /food:
    x-swagger-router-controller: food
    post:
      tags:
        - food
      operationId: newFood
      summary: post a meal picture
      description: Upload a photo of a meal for the logged user
      security:
        - Bearer: []
      x-security-scopes:
        - patient
      consumes:
        - multipart/form-data
      produces: 
        - application/json
      parameters:

        - name: image
          in: formData
          type: file
          required: true
      responses:
        "201":
          description: "Success"
          schema:
            $ref: "#/definitions/NewFoodResponse"
        "403":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: "User not found"
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /food/history/{user}:
    x-swagger-router-controller: food
    get:
      tags:
        - food
      operationId: getFoods
      summary: Retrieve food history
      description: Retrieve the food history of the  user
      security:
        - Bearer: []
      x-security-scopes:
        - patient
      consumes:
        - application/json
      produces: 
        - application/json
      parameters:
        - description: Provide a header to specify the language.
          in: header
          name: Accept-Language
          required: false
          type: string
        - name: user
          in: path
          description: The nickname of the user to retrieve food history
          type: string
          required: true
        - name: next
          in: query
          description: the id for retrieve the next page of results
          type: number
          required: false
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/FoodHistoryResponse"
        "403":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: "User not found"
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /food/images/{size}/{image}:
    x-swagger-router-controller: food
    get:
      tags:
        - food
      operationId: getFoodImageSize
      summary: Retrieve food history
      description: Retrieve the food history of the  user
      security:
        - Bearer: []
        - ApiKey: []

      x-security-scopes:
        - patient
        - admin
      consumes:
        - application/json
      produces: 
        - image/jpeg
      parameters:
        - name: size
          in: path
          description: The size of the image big medium small
          type: string
          enum: [small, medium, big]
          required: true
        - name: image
          in: path
          description: The id of the image
          type: string
          required: true
      responses:
        "200":
          description: "Success"   
        "403":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: "User not found"
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /food/images/{image}:
    x-swagger-router-controller: food
    get:
      tags:
        - food
      operationId: getFoodImage
      summary: Retrieve food history
      description: Retrieve the food history of the  user
      security:
        - Bearer: []
        - ApiKey: []

      x-security-scopes:
        - patient
      consumes:
        - application/json
      produces: 
        - image/jpeg
      parameters:
        - name: image
          in: path
          description: The id of the image
          type: string
          required: true
      responses:
        "200":
          description: "Success"   
        "403":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: "User not found"
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /food/name:
    x-swagger-router-controller: food
    post:
      tags:
        - food
      summary: "Create food  by name role : admin"
      description: Create fod by name
      operationId: newFoodName
      security:
        - Bearer: []
      x-security-scopes:
        - admin
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name:  ingredient
          in: body
          description: The food JSON you want to POST
          schema:
            $ref: "#/definitions/FoodNameObject"
          required: true
      responses:
        "201":
          description: food created
          schema:
            $ref: "#/definitions/FoodNameResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/BadRequestResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Ingredient Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /food/names/untranslated:
    x-swagger-router-controller: food
    get:
      tags:
        - food
      operationId: getUntranslatedFoods
      summary: "Retrieve untranslated foods roles: admin"
      description: Retrieve the untranslated foods 
      security:
        - Bearer: []
      x-security-scopes:
        - admin
      consumes:
        - application/json
      produces: 
        - application/json
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/FoodUntranslatedResponse"
        "403":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: "User not found"
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /food/names/{id}:
    x-swagger-router-controller: food
    get:
      tags:
        - food
      summary: "Get food names ingredients and composition by id roles: admin"
      description: get the name ingredients and composition of a food by id
      operationId: getFoodName
      security:
        - Bearer: []
        - ApiKey: []
      x-security-scopes:
        - admin
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the food to retrieve
          type: string
          required: true
      responses:
        "200":
          description: Food retrieved
          schema:
            $ref: "#/definitions/FoodNameResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Ingredient Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
    put:
      tags:
        - food
      summary: "Update food  roles: admin"
      description: Update the name of a food
      operationId: updateFoodName
      security:
        - Bearer: []
      x-security-scopes:
        - admin
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the food to update
          type: string
          required: true
        - name:  food
          in: body
          description: The food  JSON you want to POST
          schema:
            $ref: "#/definitions/FoodNameObject"
          required: true
      responses:
        "200":
          description: food updated
          schema:
            $ref: "#/definitions/ServerResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/BadRequestResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Ingredient Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /food/verify/{id}:
    x-swagger-router-controller: food
    put:
      tags:
        - food
      summary: Verify food
      description: set recognised to true
      operationId: verifyFood
      security:
        - Bearer: []
      x-security-scopes:
        - patient
        - admin
      consumes:
        - application/json
        - application/octet-stream  

      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the food to veirfy
          type: string
          required: true

      responses:
        "200":
          description: User Updated
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ServerResponse"
        "404":
          description: Food Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
        # responses may fall through to errors
  /food/unverified:
    x-swagger-router-controller: food
    get:
      tags:
        - food
      operationId: getUnverifiedFoods
      summary: Retrieve food history
      description: Retrieve the food history of the  user
      security:
        - Bearer: []
      x-security-scopes:
        - admin
      consumes:
        - application/json
      produces: 
        - application/json
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/FoodUnverifiedResponse"
        "403":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: "User not found"
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /food/{id}:
    x-swagger-router-controller: food
    get:
      tags:
        - food
      operationId: getFoodById
      summary: Retrieve users food by Id
      description: Retrieve the food history of the  user
      security:
        - Bearer: []
      x-security-scopes:
        - patient
        - admin
      consumes:
        - application/json
      produces: 
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the food
          type: string
          required: true
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/FoodResponse"   
        "403":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: "Food not found"
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /nutrients/types:
    x-swagger-router-controller: nutrient
    get:
      tags:
        - nutrients
      operationId: getNutrientTypes
      summary: get nutrients
      description: Retrive the list of nutrients and its ids
      security:
        - Bearer: []
        - ApiKey: []
      x-security-scopes:
        - patient
        - admin
        - profesional
      consumes:
        - application
      produces: 
        - application/json
      parameters:
        - description: Provide a header to specify the language.
          in: header
          name: Accept-Language
          required: false
          type: string  

      responses:
        "201":
          description: "Success"
          schema:
            $ref: "#/definitions/NutrientsTypeResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/BadRequestResponse"
        "403":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: "User not found"
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
...
definitions:
...
  NotFoundResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Error: Resource Not found"
      statusCode:
        type: number
        example: 404
...
  RuleResponse:
    required:
      - message
      - statusCode
      - rule
    properties:
      message:
        type: string
        example: "success"
      statusCode:
        type: number
        example: 200
      rule:
          $ref: "#/definitions/RuleRetrievedObject"
  ServerResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Success"
      statusCode:
        type: number
        example: 200
  NewFoodResponse:
    required:
      - message
      - statusCode
      - food_id
    properties:
      message:
        type: string
        example: "Success"
      statusCode:
        type: number
        example: 200
      food_id:
        type: string
        example: "aef425123farefe12asdffa34adea3"
...
  AccessDeniedResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Access Denied"
      statusCode:
        type: number
        example: 403
...
  ServerErrorResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Error: Internal Server Error"
      statusCode:
        type: number
        example: 500